requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.2
django-crispy-forms==2.0
Pillow==9.0.1
sqlparse==0.4.4
typing_extensions==4.7.0
